Add hook to update pod data
