+++ control-station/src/pages/VehiclePage/VehiclePage.tsx
@@ -3,20 +3,11 @@ import { Pagination } from 'components/Pagination/Pagination';
 import { PageWrapper } from 'pages/PageWrapper/PageWrapper';
 import { Outlet } from 'react-router-dom';
 import { useEmergencyOrders } from 'hooks/useEmergencyOrders';
-import { useMeasurementsStore, usePodDataStore, useSubscribe } from 'common';
+import { usePodDataUpdate } from 'hooks/usePodDataUpdate';
 
 export const VehiclePage = () => {
+    usePodDataUpdate();
     useEmergencyOrders();
-
-    const updatePodData = usePodDataStore((state) => state.updatePodData);
-    const updateMeasurements = useMeasurementsStore(
-        (state) => state.updateMeasurements
-    );
-
-    useSubscribe('podData/update', (update) => {
-        updatePodData(update);
-        updateMeasurements(update);
-    });
 
     return (
         <PageWrapper title="Vehicle">
